+++ crates/fluvio-sc-schema/src/lib.rs
@@ -85,12 +85,13 @@ mod admin {
 
     use std::{fmt::Debug};
 
-    use anyhow::Result;
+    use anyhow::{anyhow, Result};
 
     use fluvio_protocol::{Encoder, Decoder, Version};
     use fluvio_controlplane_metadata::{store::MetadataStoreObject};
 
     use crate::objects::Metadata;
+    use crate::objects::classic::ClassicCreatableAdminSpec;
 
     use super::core::{Spec};
 
@@ -114,9 +115,7 @@ mod admin {
     }
 
     /// Not every Admin Object can be created directly
-    pub trait CreatableAdminSpec: Spec + Encoder + Decoder {
-        const CREATE_TYPE: u8;
-    }
+    pub trait CreatableAdminSpec: ClassicCreatableAdminSpec + Spec + Encoder + Decoder {}
 
     pub trait DeletableAdminSpec: Spec + Encoder + Decoder {
         type DeleteKey: Encoder + Decoder + Debug + Default;
