+++ crates/fluvio-sc-schema/src/spg/mod.rs
@@ -2,14 +2,12 @@ pub use fluvio_controlplane_metadata::spg::*;
 
 mod convert {
 
-    use crate::{AdminSpec, CreatableAdminSpec, DeletableAdminSpec};
+    use crate::{AdminSpec, DeletableAdminSpec, CreatableAdminSpec};
     use super::SpuGroupSpec;
 
     impl AdminSpec for SpuGroupSpec {}
 
-    impl CreatableAdminSpec for SpuGroupSpec {
-        const CREATE_TYPE: u8 = 2;
-    }
+    impl CreatableAdminSpec for SpuGroupSpec {}
 
     impl DeletableAdminSpec for SpuGroupSpec {
         type DeleteKey = String;
