+++ crates/fluvio-sc-schema/src/tableformat/mod.rs
@@ -2,16 +2,14 @@ pub use fluvio_controlplane_metadata::tableformat::*;
 
 mod convert {
 
-    use crate::{CreatableAdminSpec, DeletableAdminSpec};
+    use crate::{DeletableAdminSpec, CreatableAdminSpec};
 
     use crate::{AdminSpec};
     use super::TableFormatSpec;
 
     impl AdminSpec for TableFormatSpec {}
 
-    impl CreatableAdminSpec for TableFormatSpec {
-        const CREATE_TYPE: u8 = 5;
-    }
+    impl CreatableAdminSpec for TableFormatSpec {}
 
     impl DeletableAdminSpec for TableFormatSpec {
         type DeleteKey = String;
